Parameters dialog: Pass most keys through to REAPER's main window.
Previously, we only passed specific keys through to the main window.
Now, we do the inverse, capturing only specific keys for the Parameters dialog and passing the rest through to the main window.
The combo box and text boxes are still exceptions, where we only pass specific keys to the main window to ensure the user can interact with the control properly.
Practically speaking, this means users can access many more main window actions when focused on the slider or buttons.